requirements.txt
flask
openai-whisper
librosa
tensorflow==2.13.0
scikit-learn
pandas
numpy
